+++ README.md
@@ -1,6 +1,2 @@
 # TTMS-Coding-Practice
 This is a project about Theatre Ticketing Management System(TTMS) written in C. This is a coding practice.
-
-## V1.0
-
-TTMS V1.0 was released at 4/7/2019. The whole program is basically self-explanatory. The only logical difference you should be aware of is that there's no sense of "time", meaning that when you make a purchase, you are finished with the movie, but when you return a ticket, the ticket will be like as if it hasn't been purchased before.
