+++ TTMS/TTMS/TTMS.vcxproj.filters
@@ -63,9 +63,6 @@
     <ClCompile Include="ticketRelated.c">
       <Filter>源文件\basic</Filter>
     </ClCompile>
-    <ClCompile Include="dataAnalysis.c">
-      <Filter>源文件\basic</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="usedVariableTypes.h">
