+++ TTMS/TTMS/ticketRelated.c
@@ -90,43 +90,3 @@ struct ticketHistory* ticketHistoryCheckByID(long long int inputID)
 		return NULL;
 	}
 }
-
-struct ticketHistory* unsoldOrWatchedTicketHistoryCheckByStudioAndPlaceAndDay(struct ticketHistory* ticketHistoryStart, struct studio* studioSwap, int today, int whichLine, int whichColumn)
-{
-	struct ticketHistory* ticketHistorySwap;
-	for (long long int whichTicket = 1; whichTicket <= howManyTicketHistory(ticketHistoryStart); whichTicket++)
-	{
-		ticketHistorySwap = ticketHistoryCheckByWhichOne(ticketHistoryStart, whichTicket);
-		if (ticketHistorySwap->studioID == studioSwap->ID && ticketHistorySwap->shouldWatchDay == today && ticketHistorySwap->whichColumn == whichColumn && ticketHistorySwap->whichLine == whichLine && (ticketHistorySwap->status == 0 || ticketHistorySwap->status == 1))
-		{
-			return ticketHistorySwap;
-		}
-	}
-	return NULL;
-}
-
-void printTicket(struct ticketHistory* ticketHistorySwap)
-{
-	printf("Ticket ID: %lld\n", ticketHistorySwap->ticketID);
-	struct movie* movieSwap = movieCheckByID(ticketHistorySwap->movieID);
-	printf("Movie: %s\n", movieSwap->name);
-	printf("Price: %lld\n", movieSwap->price);
-	printf("Studio ID: %lld\n", ticketHistorySwap->studioID);
-	printf("Watch day: %lld\n", ticketHistorySwap->shouldWatchDay);
-	printf("WhichLine: %d\n", ticketHistorySwap->whichLine);
-	printf("WhichLine: %d\n", ticketHistorySwap->whichColumn);
-	printf("Status: ");
-	if (ticketHistorySwap->status == 0)
-	{
-		printf("Unsold\n");
-	}
-	else if (ticketHistorySwap->status == 1)
-	{
-		printf("Sold/Watched\n");
-	}
-	else if (ticketHistorySwap->status == 2)
-	{
-		printf("Error\n");
-	}
-	printf("\n");
-}
